Remove state from Network module
Framework Network module remove getUniqueOutboundConnectedPeersCount()
